Move AddressType to common.py
