+++ hwilib/commands.py
@@ -5,7 +5,7 @@
 import importlib
 import platform
 
-from .serializations import AddressType, PSBT
+from .serializations import PSBT
 from ._base58 import xpub_to_pub_hex
 from .key import (
     H_,
@@ -31,7 +31,10 @@ from .descriptor import (
     WSHDescriptor,
 )
 from .devices import __all__ as all_devs
-from .common import Chain
+from .common import (
+    AddressType,
+    Chain,
+)
 from .hwwclient import HardwareWalletClient
 
 from itertools import count
