+++ hwilib/serializations.py
@@ -24,7 +24,6 @@ import hashlib
 import copy
 import base64
 
-from enum import Enum
 from io import BytesIO, BufferedReader
 from typing import (
     Dict,
@@ -35,7 +34,6 @@ from typing import (
     Sequence,
     Tuple,
     TypeVar,
-    Union,
     Callable,
 )
 from typing_extensions import Protocol
@@ -64,24 +62,6 @@ def hash256(s: bytes) -> bytes:
 def hash160(s: bytes) -> bytes:
     return ripemd160(sha256(s))
 
-
-class AddressType(Enum):
-    PKH = 1
-    WPKH = 2
-    SH_WPKH = 3
-
-    def __str__(self) -> str:
-        return self.name.lower()
-
-    def __repr__(self) -> str:
-        return str(self)
-
-    @staticmethod
-    def argparse(s: str) -> Union['AddressType', str]:
-        try:
-            return AddressType[s.upper()]
-        except KeyError:
-            return s
 
 # Serialization/deserialization tools
 def ser_compact_size(size: int) -> bytes:
